+++ src/main/java/com/team/arium/domain/Faq_Post.java
@@ -17,7 +17,7 @@ public class Faq_Post {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "faq_id")
-    private Long faqId;
+    private Integer faqId;
     
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "ntc_id", nullable = false)
